requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.40.24
botocore==1.40.24
certifi==2025.11.12
charset-normalizer==3.4.4
Deprecated==1.3.1
Django==6.0
frozenlist==1.8.0
future==1.0.0
idna==3.11
jmespath==1.0.1
line-bot-sdk==3.21.0
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==2.0.1
yarl==1.22.0
gunicorn==21.2.0
